+++ src/js/template/bar.js
@@ -1,0 +1,19 @@
+'use strict';
+
+// Default branding bar template
+var template = '' +
+'  <div class="branding-bar_container">' +
+'    <div class="branding-bar_links">' +
+'      <a class="social" href="https://www.facebook.com/sunlightfoundation"><span class="sficon-facebook"></span></a>' +
+'      <a class="social" href="https://twitter.com/sunfoundation"><span class="sficon-twitter"></span></a>' +
+'      <a class="social" href="https://plus.google.com/+sunlightfoundation"><span class="sficon-google-plus"></span></a>' +
+'      <a class="branding-bar_trigger" data-bb-toggle=".bb_wrapper" href="https://sunlightfoundation.com/about/">About Sunlight Foundation</a>' +
+'    </div>' +
+'    <div class="branding-bar_logo">' +
+'      <span class="branding-bar_productof">a product of </span>' +
+'      <a class="branding-bar_sunlight-logo" href="https://www.sunlightfoundation.com">Sunlight Foundation</a>' +
+'    </div>' +
+'  </div>' +
+'';
+
+module.exports = template;
